cook

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: fakman <student@42.tr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/08 22:26:49 by fakman            #+#    #+#             */
-/*   Updated: 2022/12/16 11:26:35 by fakman           ###   ########.fr       */
+/*   Updated: 2022/12/23 11:06:06 by fakman           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	int	counter;
+	size_t	counter;
 
 	counter = 0;
 	while (counter < n)
